requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
reportlab==4.0.7
openpyxl==3.1.2
gunicorn==21.2.0
mongoengine==0.28.2
pymongo==4.6.1
python-dotenv==1.0.0
dnspython==2.5.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
